+++ Flow/test/utopia/flow/test/collection/MatrixTest.scala
@@ -13,16 +13,18 @@ object MatrixTest extends App
 	val width = 8
 	val matrix = Matrix.fill(Pair.twice(width)) { case Pair(x, y) => y * width + x }
 	
-	def printMatrix(m: Matrix[_]) = {
-		println("\n---")
+	def printMatrix(m: Matrix[_], name: String) = {
+		println(s"\n--- $name ---")
 		m.rows.foreach { row => println(row.mkString(", ")) }
 		println("---")
 	}
 	
-	printMatrix(matrix)
-	printMatrix(matrix.viewBetween(Pair(Pair.twice(1), Pair.twice(3))))
+	printMatrix(matrix, "all")
+	printMatrix(matrix.viewBetween(Pair(Pair.twice(1), Pair.twice(3))), "1-3")
 	
 	val cropped = matrix.view(Pair(7, 3).map { p => NumericSpan(p - 1, p + 1) })
-	printMatrix(cropped)
+	printMatrix(cropped, "cropped sub-region")
 	println(cropped.iterator.mkString(", "))
+	
+	printMatrix(matrix.crop(Pair(1, 2)), "cropped")
 }